backend/requirements.txt
fastapi
uvicorn[standard]

pydantic
pydantic-settings
python-dotenv

databases
asyncpg

python-jose
firebase-admin

apscheduler
httpx
requests

pandas
numpy

matplotlib
seaborn

scikit-learn
spacy

openai

ckanapi
datagovindia

python-multipart
aiofiles